+++ sockeye/translate.py
@@ -69,12 +69,14 @@ def run_translate(args: argparse.Namespace):
 
     device = init_device(args, logger)
     logger.info(f"Translate Device: {device}")
+
     models, source_vocabs, target_vocabs = load_models(device=device,
                                                        model_folders=args.models,
                                                        checkpoints=args.checkpoints,
                                                        dtype=args.dtype,
                                                        clamp_to_dtype=args.clamp_to_dtype,
-                                                       inference_only=True)
+                                                       inference_only=True,
+                                                       knn_index=args.knn_index)
 
     restrict_lexicon = None  # type: Optional[Union[RestrictLexicon, Dict[str, RestrictLexicon]]]
     if args.restrict_lexicon is not None:
@@ -134,6 +136,7 @@ def run_translate(args: argparse.Namespace):
                                       sample=args.sample,
                                       output_scores=output_handler.reports_score(),
                                       constant_length_ratio=constant_length_ratio,
+                                      knn_lambda=args.knn_lambda,
                                       max_output_length_num_stds=args.max_output_length_num_stds,
                                       max_input_length=args.max_input_length,
                                       max_output_length=args.max_output_length,